requirements.txt
transformers>=4.40.0
geoopt>=0.5.0
scikit-learn>=1.3.0
PyYAML>=6.0
matplotlib>=3.7.0
Pillow>=10.0.0
safetensors>=0.4.0
open_clip_torch>=2.24.0
scipy>=1.11.0
tabulate>=0.9.0
tqdm>=4.66.0
torch>=2.0.0
torchvision>=0.15.0
einops>=0.7.0
timm>=0.9.0
fvcore>=0.1.5.post20221221
iopath>=0.1.10
simplejson>=3.19.2
